pyproject.toml
[tool.poetry]
name = "ml-pipeline"
version = "0.1.0"
description = "Config-driven ML pipeline for power consumption analysis"
authors = ["kalgoritmi"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.0"
scikit-learn = "^1.3.0"
pydantic = "^2.0.0"
pyyaml = "^6.0"

[tool.poetry.scripts]
ml-pipeline = "src.pipeline:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
